Updated debug.c with shell.c

+++ debug.c
@@ -15,7 +15,12 @@ int background_proc[MAX_BG_PROCESS];
 int foreground_proc[MAX_FG_PROCESS];
 int interrupt;
 
-// Splits the string by space and returns the array of tokens
+/**
+ * @fn tokenize
+ * @param[in] line
+ * @return tokens
+ * @brief Split line into tokens
+ */
 char **tokenize(char *line) {
   char **tokens = (char **)malloc(MAX_NUM_TOKENS * sizeof(char *));
   char *token = (char *)malloc(MAX_TOKEN_SIZE * sizeof(char));
@@ -42,9 +47,16 @@ char **tokenize(char *line) {
   return tokens;
 }
 
+/**
+ * @fn background
+ * @param[in] tokens
+ * @brief Run the command by forking and calling executable (except cd).
+ *        Don't wait for it to end (run as background process)
+ */
 void background(char **tokens) {
   int i;
 
+  // Check availability of background process and set i accordingly
   for (i = 0; i < MAX_BG_PROCESS; i++) {
     if (background_proc[i] == -1) {
       break;
@@ -54,31 +66,58 @@ void background(char **tokens) {
     printf("Shell: Can't handle more background processes\n");
     return;
   }
+
+  // Fork to run the the command
   int ret = fork();
   if (ret < 0) {
     printf("Shell: Error while calling fork\n");
-  } else if (ret == 0) { // Child process
+  } else if (ret == 0) {
+    // Child process
+    // Assign a different process group id
     setpgid(0, 0);
+
     if (tokens[0] == NULL) {
+      // Nothing to do
       printf("Shell: Nothing to do\n");
     } else if (!strcmp(tokens[0], "cd")) {
-      printf("Shell: cd in background doesn't make sense\n");
+      // Special case cd
+      // Doesn't make sense to run cd in background
+      // Even then, check for format of argument
+      if (tokens[1] == NULL || tokens[2] != NULL)
+        printf("Shell: Incorrect command\n");
+      else {
+        int l = chdir(tokens[1]);
+        if (l == -1) {
+          printf("Shell: Directory not found\n");
+        }
+      }
     } else {
+      // Load and run the executable
       int p = execvp(tokens[0], tokens);
       if (p == -1) {
         printf("Shell: Incorrect command\n");
       }
     }
     exit(0);
-  } else { // ret > 0, Parent process with ret as Child PID
+  } else { // ret > 0
+    // Parent process with ret as Child PID
     background_proc[i] = ret;
   }
 }
 
+/**
+ * @fn normal
+ * @param[in] tokens
+ * @brief Run the command by forking and calling executable (except cd).
+ *        Wait for it to end (run as foreground process)
+ */
 void normal(char **tokens) {
   if (tokens[0] == NULL) {
+    // Nothing to do
     printf("Shell: Nothing to do\n");
   } else if (!strcmp(tokens[0], "cd")) {
+    // Special case cd
+    // Check for format of argument
     // TODO: Add "cd -" compatibility for fun
     if (tokens[1] == NULL || tokens[2] != NULL)
       printf("Shell: Incorrect command\n");
@@ -89,16 +128,21 @@ void normal(char **tokens) {
       }
     }
   } else {
+    // Fork to run the the command
     int ret = fork();
     if (ret < 0) {
       printf("Shell: Error while calling fork\n");
-    } else if (ret == 0) { // Child process
+    } else if (ret == 0) {
+      // Child process
+      // Load and run the executable
       int p = execvp(tokens[0], tokens);
       if (p == -1) {
         printf("Shell: Incorrect command\n");
       }
       exit(0);
-    } else { // ret > 0, Parent process with ret as Child PID
+    } else { // ret > 0
+      // Parent process with ret as Child PID
+      // Wait for the child process to terminate then reap it
       int k = waitpid(ret, NULL, 0);
       if (k == -1) {
         printf("Shell: Error while calling waitpid\n");
@@ -107,10 +151,16 @@ void normal(char **tokens) {
   }
 }
 
+/**
+ * @fn run
+ * @param[in] tokens
+ * @brief Detect background process and accordingly run the command
+ */
 void run(char **tokens) {
   int i;
   int bg = 0;
 
+  // Don't run if interrupt is set to 1
   if (interrupt == 1) {
     printf("Cancelling: ");
     for (i = 0; tokens[i] != NULL; i++) {
@@ -119,13 +169,13 @@ void run(char **tokens) {
     printf("\n");
     return;
   }
-
   printf("Running: ");
   for (i = 0; tokens[i] != NULL; i++) {
     printf("%s ", tokens[i]);
   }
   printf("\n");
 
+  // Check if it is background (ends with "&") or not
   for (i = 0; tokens[i] != NULL; ++i) {
     if (!strcmp(tokens[i], "&") && tokens[i + 1] == NULL) {
       tokens[i] = NULL;
@@ -134,6 +184,7 @@ void run(char **tokens) {
     }
   }
 
+  // Separately run background process
   if (bg) {
     background(tokens);
   } else {
@@ -141,30 +192,46 @@ void run(char **tokens) {
   }
 }
 
+/**
+ * @fn series
+ * @param[in] tokens
+ * @brief Splits the token into components based on "&&".
+ *        Run them one after another in foreground
+ */
 void series(char **tokens) {
   char **ptokens = (char **)malloc(MAX_NUM_TOKENS * sizeof(char *));
   int i, j;
 
+  // Split into components
   j = 0;
   for (i = 0; tokens[i] != NULL; ++i) {
     if (strcmp(tokens[i], "&&")) {
       ptokens[j] = tokens[i];
       ++j;
     } else {
+      // If encountered a "&&", segment till now is run and waited for
       ptokens[j] = NULL;
       run(ptokens);
       j = 0;
     }
   }
+  // Same for last segment
   ptokens[j] = NULL;
   run(ptokens);
 
+  // Free the allocated memory
   free(ptokens);
 }
 
+/**
+ * @fn parallel
+ * @param[in] tokens
+ * @brief Run the command on a new shell as a foreground process
+ */
 void parallel(char **tokens) {
   int i;
 
+  // Check availability of foreground process and set i accordingly
   for (i = 0; i < MAX_FG_PROCESS; i++) {
     if (foreground_proc[i] == -1) {
       break;
@@ -174,36 +241,50 @@ void parallel(char **tokens) {
     printf("Shell: Can't handle more foreground processes\n");
     return;
   }
+
+  // Fork another shell to run the command
   int ret = fork();
   if (ret < 0) {
     printf("Shell: Error while calling fork\n");
-  } else if (ret == 0) { // Child process
+  } else if (ret == 0) {
+    // Child process
     // signal(SIGINT, SIG_DFL);
     series(tokens);
     exit(0);
-  } else { // ret > 0, Parent process with ret as Child PID
+  } else { // ret > 0
+    // Parent process with ret as Child PID
     foreground_proc[i] = ret;
   }
 }
 
+/**
+ * @fn work
+ * @param[in] tokens
+ * @brief Splits the token into components based on "&&&".
+ *        Run them parallelly in foreground
+ */
 void work(char **tokens) {
   char **ptokens = (char **)malloc(MAX_NUM_TOKENS * sizeof(char *));
   int i, j;
 
+  // Split into components
   j = 0;
   for (i = 0; tokens[i] != NULL; ++i) {
     if (strcmp(tokens[i], "&&&")) {
       ptokens[j] = tokens[i];
       ++j;
     } else {
+      // If encountered a "&&&", segment till now is run on a new shell
       ptokens[j] = NULL;
       parallel(ptokens);
       j = 0;
     }
   }
+  // Last segment is run on the current shell itself
   ptokens[j] = NULL;
   series(ptokens);
 
+  // Wait for all foreground processes to end
   for (i = 0; i < MAX_FG_PROCESS; ++i) {
     if (foreground_proc[i] > 0) {
       int k = waitpid(foreground_proc[i], NULL, 0);
@@ -216,12 +297,20 @@ void work(char **tokens) {
     }
   }
 
+  // Free the allocated memory
   free(ptokens);
 }
 
+/**
+ * @fn handle_sig
+ * @param[in] sig
+ * @brief SIGINT handler
+ *        Does nothing, just sets interrupt to 1 (no new command will run)
+ */
 void handle_sig(int sig) {
   printf("\n");
   interrupt = 1;
+  // SIGINT passed to children in same process group as well
 }
 
 int main(int argc, char *argv[]) {
@@ -229,8 +318,10 @@ int main(int argc, char *argv[]) {
   char **tokens;
   int i;
 
+  // SIGINT handler added
   signal(SIGINT, handle_sig);
 
+  // Initialize list of background and foreground processes
   for (i = 0; i < MAX_BG_PROCESS; ++i) {
     background_proc[i] = -1;
   }
@@ -239,6 +330,7 @@ int main(int argc, char *argv[]) {
   }
 
   while (1) {
+    // Scan the line
     bzero(line, sizeof(line));
     printf("$ ");
     scanf("%[^\n]", line);
@@ -246,6 +338,7 @@ int main(int argc, char *argv[]) {
 
     // printf("Command entered: %s (remove this debug output later)\n", line);
 
+    // Reap background child processes which have ended
     for (i = 0; i < MAX_BG_PROCESS; i++) {
       if (background_proc[i] > 0) {
         int k = waitpid(background_proc[i], NULL, WNOHANG);
@@ -258,36 +351,48 @@ int main(int argc, char *argv[]) {
       }
     }
 
-    line[strlen(line)] = '\n'; // terminate with new line
-    tokens = tokenize(line);   // convert line to tokens
+    // Break the line into tokens
+    line[strlen(line)] = '\n';
+    tokens = tokenize(line);
+
     if (tokens[0] == NULL) {
+      // Nothing to do
       printf("Shell: Nothing to do\n");
     } else if (!strcmp(tokens[0], "exit")) {
-      for (i = 0; i < MAX_BG_PROCESS; i++) {
-        if (background_proc[i] > -1) {
-          kill(background_proc[i], SIGKILL);
-          printf("Shell: Background process [%i] killed\n", background_proc[i]);
-          background_proc[i] = -1;
+      if (tokens[1] == NULL) {
+        // Kill background processes before exit
+        for (i = 0; i < MAX_BG_PROCESS; i++) {
+          if (background_proc[i] > -1) {
+            kill(background_proc[i], SIGKILL);
+            printf("Shell: Background process [%i] killed\n",
+                   background_proc[i]);
+            background_proc[i] = -1;
+          }
         }
-      }
 
-      // Freeing the allocated memory
-      for (i = 0; tokens[i] != NULL; i++) {
-        free(tokens[i]);
-      }
-      free(tokens);
+        // Free the allocated memory
+        for (i = 0; tokens[i] != NULL; i++) {
+          free(tokens[i]);
+        }
+        free(tokens);
 
-      return 0;
+        // Just exit
+        return 0;
+      } else {
+        printf("Shell: exit command doesn't have any arguments\n");
+      }
     } else {
+      // Set interrupt to 0 (new command will run)
       interrupt = 0;
-      work(tokens); // work on the tokens
+      // Work on the tokens
+      work(tokens);
     }
 
     // for (i = 0; tokens[i] != NULL; i++) {
     //   printf("found token %s (remove this debug output later)\n", tokens[i]);
     // }
 
-    // Freeing the allocated memory
+    // Free the allocated memory
     for (i = 0; tokens[i] != NULL; i++) {
       free(tokens[i]);
     }
